Sensores_IoT7/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pyjwt==2.8.0
scikit-learn==1.3.0
numpy==1.24.3
requests==2.31.0
paho-mqtt==1.6.1